requirements.txt
Django
django-import_export
djangorestframework
django-filter
djangorestframework-word-filter
django-cors-headers